début implémentation password
incomplet tests sont laissés

+++ www/ProjetPreTPI/src/php/API/deleteAllRecord.php
@@ -2,7 +2,15 @@
     require('./src/php/Application/DBHelper.php');
 
     $database = new DBHelper;
-    $database->deleteAllRecord();
+    
+    //Password authentification
+    
+    
+
+    if ($userPass == $password) {
+        $database->deleteAllRecord();
+    }
+
 
     header("Location: /");
 ?>
